flake: inputs: remove flake-parts
`flake-parts` might be useful for large projects with custom flake
outputs, or where flake outputs are composed in unique ways. Alas, for
this flake, that is not the case, so `flake-parts` is just introducing
complexity without providing value achieving critical-mass.

This commit will be followed by other improvements. For now, removing
flake-parts necessarily complicates the output expressions, since the
machinery is now expressed inline through `eachSystem` and `pkgsFor`.
That is expected and will be remedied.